Actually make JAX import into WORKSPACE work.
The current import was non-functional, and that only remained unnoticed
because there is no BUILD dependency to JAX, so JAX isn't actually
built. (The CI tests install JAX through pip, so the JAX dependency
isn't needed there.)

The fix consists in moving the configuration from JAX's WORKSPACE file
(which is emptied when imported via `http_archive`) to the WORKSPACE
file of this project (plus fixing how its dependency `ducc` is loaded).
Now adding `@jax//jax` to some of the build targets actually (and
successfully) builds JAX.