Handle simple "else" branches in affine simplification
Vendor parts of the affine domain extraction in order to add support for "else"
branches of `affine.if`. This is currently restricted to conditions the
inversion of which has only one disjunct. This is due to the ISL conversion
logic in affine simplification not supporting `PresburgerSet` with disjunctions
but only single `IntegerPolyhedron`.

Fix a logic bug identifing the else region, the `op` value was shadowed by the
range-for induction variable making the condition for detecting the "else"
branch never true.

This enables us to raise `memref.load` depending on values yielded from
`affine.if` to StableHLO.