Handle simple "else" branches in affine simplification (#527)
* lift memref.load depending on scf.if into that

When a load depends on the conditional value produced by an `scf.if`,
lift that load into both conditional branches when possible (no
interferring side-effecting operations). This may allow us to turn such
loads into affine loads and further raise them.

Additional patterns are needed to fold arithmetic operations into affine
forms and further use those to construct and affine load.

* Handle simple "else" branches in affine simplification

Vendor parts of the affine domain extraction in order to add support for "else"
branches of `affine.if`. This is currently restricted to conditions the
inversion of which has only one disjunct. This is due to the ISL conversion
logic in affine simplification not supporting `PresburgerSet` with disjunctions
but only single `IntegerPolyhedron`.

Fix a logic bug identifing the else region, the `op` value was shadowed by the
range-for induction variable making the condition for detecting the "else"
branch never true.

This enables us to raise `memref.load` depending on values yielded from
`affine.if` to StableHLO.

---------

Co-authored-by: Mosè Giordano <765740+giordano@users.noreply.github.com>