Add a dialect-specific canonicalization pass (#1172)
This only applies canonicalization patterns for ops from the dialects listed in
pass options. Normally those ops list patterns for themselves, but it is not a
strict requirement.

This offers significant performance bump compared to general canonicalization
that collects patterns from all loaded dialects, many of which may not be
present at certain stages of compilation. This wastes cycles trying to match
patterns that would never apply.